requirements.txt
altgraph==0.17.4
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
executing==2.0.1
future==1.0.0
ipython==8.26.0
iso8601==2.1.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pefile==2023.2.7
prompt_toolkit==3.0.47
pure-eval==0.2.2
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyserial==3.5
pywin32-ctypes==0.2.2
PyYAML==6.0.1
setuptools==70.1.1
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
